+++ src/components/Header.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from "react-router-dom";
 import StyledHeader from '@styles/components/Header.styled';
 
-const Header = (): JSX.Element =>
+const Header: React.FC = (): JSX.Element =>
   <StyledHeader>
     <h6><Link to="/dashboard">TIMEWARD</Link></h6>
     <nav>