+++ src/components/PieChart.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import StyledPieChart from '@styles/components/PieChart.styled';
 
-const PieChart = (): JSX.Element => <StyledPieChart />;
+const PieChart: React.FC = (): JSX.Element => <StyledPieChart />;
 
 export default PieChart;