+++ src/pages/Login.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Alignment from '@components/Alignment';
 import Form from '@components/Form';
 
-const Login = (): JSX.Element =>
+const Login: React.FC = (): JSX.Element =>
   <Alignment vertical horizontal>
     <Form>
       <Form.Input />